% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_pollutionRoseServer.R
\name{arrow_pollutionRoseServer}
\alias{arrow_pollutionRoseServer}
\title{Pollution Rose Server}
\usage{
arrow_pollutionRoseServer(
  id,
  data_rct,
  data_wind_rct,
  wd_col,
  param_col,
  measurement_col,
  key_col,
  unit_col,
  site_col,
  param_rct,
  brush = reactive({     NULL }),
  title = NULL
)
}
\arguments{
\item{id}{string ID used to link pollutionRoseUI() to pollutionRoseServer()}

\item{data_rct}{reactive Concentration data. Should have
columns for site, parameter, unit
and a "key" used in joining wind to concentration
data (typically a timestamp).}

\item{data_wind_rct}{reactive Wind data. Should have column for
direction and speed, and the same
key_col and site_col as data_rct. Speed can be
filled with NA if unavailable.}

\item{wd_col}{string Name of column in wind data with direction values.}

\item{param_col}{string Name of column in 1-hr concentration data with
parameter name values.}

\item{measurement_col}{string Name of column in 1-hr concentration data with
parameter name values.}

\item{key_col}{string Name of column in 1-hr concentration and wind data with
key values (typically datetime column).}

\item{unit_col}{string Name of column in 1-hr concentration data with
unit values.}

\item{site_col}{string Name of column in 1-hr concentration data with
sitecode values.}

\item{param_rct}{reactive string Parameter to plot on windrose. Used to
filter 1-hr concentration data via the
param_col.}

\item{brush}{(optional) reactive brush Brush object with elements for x and y that
are each vectors with x/y min and x/y max values
respectively (e.g. list(x=c(0,10), y=c(-10,10)))}

\item{title}{(optional) string Title for plot. If not specified, title will be current value of param_rct.}
}
\value{

}
\description{
Pollution Rose Server
}
