% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_pollutionRoseTabServer.R
\name{arrow_pollutionRoseTabServer}
\alias{arrow_pollutionRoseTabServer}
\title{Pollution rose tab Server}
\usage{
arrow_pollutionRoseTabServer(
  id,
  ds_uri,
  datetime_col,
  site_year,
  map_data,
  wd_param_name,
  param1,
  param_col,
  sample_measurement_col,
  sample_dur_col,
  site_rct,
  site_col,
  sample_dur = "1 HOUR",
  breaks_aqi_1,
  param2 = NULL,
  param3 = NULL,
  param4 = NULL,
  unit_str_1 = NULL,
  unit_str_2 = NULL,
  unit_str_3 = NULL,
  unit_str_4 = NULL,
  title_1 = shiny::reactive({
     "PM[2.5]"
 }),
  title_2 = shiny::reactive({
     "O[3]"
 }),
  title_3 = shiny::reactive({
     "Fill title_3!"
 }),
  title_4 = shiny::reactive({
     "Fill title_4!"
 }),
  year_filter = TRUE,
  year_col = NULL,
  season_filter = TRUE,
  dow_filter = FALSE,
  hour_filter = FALSE,
  breaks_conc_1 = NULL,
  breaks_conc_2 = NULL,
  breaks_aqi_2 = NULL,
  breaks_conc_3 = NULL,
  breaks_aqi_3 = NULL,
  breaks_conc_4 = NULL,
  breaks_aqi_4 = NULL,
  aqipalette
)
}
\arguments{
\item{id}{string ID used to link pollutionRoseTabUI() to pollutionRoseTabServer()}

\item{ds_uri}{string uri to Arrow dataset, formatted as .parquet}

\item{datetime_col}{string Name of column with datetime values in posixct format}

\item{site_year}{dataframe Dataframe with distinct combinations of AQS sitecode and year}

\item{map_data}{reactive data.frame Dataframe with lat and lng columns
with the location of program sites,
to be plotted on a leaflet map.}

\item{wd_param_name}{string Parameter name for wind direction values}

\item{param1}{string Name of first pollutant parameter}

\item{param_col}{string Name of column with parameter name values}

\item{sample_measurement_col}{string Name of column with concentration values}

\item{sample_dur_col}{string Name of column with sample duration values}

\item{site_rct}{reactive string AQS sitecode for selected site}

\item{site_col}{string Column name for column with AQS sitecode values}

\item{sample_dur}{string Sample duration of data -- defaults to '1 HOUR'.}

\item{breaks_aqi_1}{vector Manual scale breaks for AQI designations for first
parameter.}

\item{param2}{string Name of optional 2nd pollutant parameter}

\item{param3}{string Name of optional 3rd pollutant parameter}

\item{param4}{string Name of optional 4th pollutant parameter}

\item{unit_str_1}{string Unit of title_1. Can be an expression.}

\item{unit_str_2}{(optional) string Unit of optional title_2. Can be an expression.}

\item{unit_str_3}{(optional) string Unit of optional title_3. Can be an expression.}

\item{unit_str_4}{(optional) string Unit of optional title_4. Can be an expression.}

\item{title_1}{reactive string Name of first parameter to include in plot title
(can be an expression if subscripts needed).}

\item{title_2}{reactive string (optional) Name of optional second parameter to include in plot title
(can be an expression if subscripts needed).}

\item{title_3}{reactive string (optional) Name of optional second parameter to include in plot title
(can be an expression if subscripts needed).}

\item{title_4}{reactive string (optional) Name of optional second parameter to include in plot title
(can be an expression if subscripts needed).}

\item{year_filter}{boolean If true, multi-select year input included and
plots may be filtered by year.}

\item{year_col}{string (optional) Column name of column with year values if year_filter specified.}

\item{season_filter}{boolean If true, multi-select season input included and
plots may be filtered by season.}

\item{dow_filter}{boolean If true, multi-select day of week input included and
plots may be filtered by day of week.}

\item{hour_filter}{boolean If true, multi-select hour input included and
plots may be filtered by hour}

\item{breaks_conc_1}{vector (optional) Manual scale breaks in legend for first
parameter. If not specified, sequence of
6 equal breaks will be created with endpoints
at the max and min concentration value.}

\item{breaks_conc_2}{(optional) Manual scale breaks in legend for optional second
parameter. If not specified, sequence of
6 equal breaks will be created with endpoints
at the max and min concentration value.}

\item{breaks_aqi_2}{(optional) vector Manual scale breaks for AQI designations for
optional second parameter.}

\item{breaks_conc_3}{vector (optional) Manual scale breaks in legend for first
parameter. If not specified, sequence of
6 equal breaks will be created with endpoints
at the max and min concentration value.}

\item{breaks_aqi_3}{vector Manual scale breaks for AQI designations for first
parameter.}

\item{breaks_conc_4}{(optional) Manual scale breaks in legend for optional second
parameter. If not specified, sequence of
6 equal breaks will be created with endpoints
at the max and min concentration value.}

\item{breaks_aqi_4}{(optional) vector Manual scale breaks for AQI designations for
optional second parameter.}

\item{aqipalette}{vector Standard color palette for aqi categories with elements:
- aqipalette$greenGood,
- aqipalette$yellowModerate,
- aqipalette$orangeUSG,
- aqipalette$redUnhealthy,
- aqipalette$purpleVeryUnhealthy,
- aqipalette$maroonHazardous}
}
\description{
Used alongside pollutionRoseTabUI() to add a collection of reactive
elements to an app, including multiselects for seasons and years, a select
for metric type (aqi or concentration), and up to 4 pollution roses for
four specified parameters.
}
