% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_timeseriesInvestigationTabServer.R
\name{arrow_timeseriesInvestigationTabServer}
\alias{arrow_timeseriesInvestigationTabServer}
\title{Timeseries Investigation Tab Server -- Arrow}
\usage{
arrow_timeseriesInvestigationTabServer(
  id,
  ds,
  selected_site_rct,
  site_dur_param_poc,
  dt_col,
  param_col,
  unit_col,
  measurement_col,
  sampledur_col,
  poc_col,
  wd_param_name,
  site_col,
  linesize = 0.75,
  pointsize = 2
)
}
\arguments{
\item{id}{string ID the connects timeseriesInvestigationTabUI() to timeseriesInvestigationTabServer()}

\item{ds}{dataset Arrow dataset object, created with arrow::open_dataset()}

\item{selected_site_rct}{reactive string Selected AQS Sitecode}

\item{site_dur_param_poc}{dataframe dataframe with unique combinations of
aqs_sitecode, sample_duration, parameter,
poc (with these column names)}

\item{dt_col}{string Name of column with datetime in posixct format}

\item{param_col}{string Name of column with parameter name}

\item{unit_col}{string name of column with units}

\item{measurement_col}{string Name of column with concentration measurements}

\item{sampledur_col}{string Name of column with sample duration values}

\item{poc_col}{string Name of column with POC values}

\item{wd_param_name}{string Parameter name for wind direction values}

\item{site_col}{string Name of column with aqs sitecodes}

\item{linesize}{numeric (Optional) size of line on timeseries, default = 0.75.}

\item{pointsize}{numeric (Optional) size of point on timeseries, default = 3.}
}
\value{

}
\description{
Used alongside timeseriesInvestigationTabUI() to add a collection of
reactive elements to the app, including multiple parameter selections,
a timeseries plot, a scatter plot, and a pollution rose. Wrapper for standalone
functions scatterServer(), timeseriesServer() and pollutionRoseServer().
Data input is an arrow dataset.
}
