% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_timeseriesInvestigationTabUI.R
\name{arrow_timeseriesInvestigationTabUI}
\alias{arrow_timeseriesInvestigationTabUI}
\title{Timeseries Investigation Tab UI}
\usage{
arrow_timeseriesInvestigationTabUI(id)
}
\arguments{
\item{id}{string ID the connects timeseriesInvestigationTabUI() to timeseriesInvestigationTabServer()}
}
\description{
Used alongside timeseriesInvestigationTabServer() to add a collection of
reactive elements to the app, including multiple parameter selections,
a timeseries plot, a scatter plot, and a pollution rose. Wrapper for standalone
functions scatterServer(), timeseriesServer() and pollutionRoseServer().
}
