% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotModServer.R
\name{boxplotModServer}
\alias{boxplotModServer}
\title{Boxplot Mod Server Function}
\usage{
boxplotModServer(
  id,
  data,
  x,
  ymin,
  lower,
  upper,
  ymax,
  median,
  fill = NULL,
  color = NULL,
  x_group = NULL,
  ann = NULL,
  ann_position_adjust = 0.4,
  x_lab = "",
  y_lab = "",
  fill_lab = "",
  color_lab = "",
  title = "",
  hline_int = NA_real_,
  coord_flip = TRUE
)
}
\arguments{
\item{id}{string ID label that links boxplotUI() to boxplotServer()}

\item{data}{reactive Reactive dataframe}

\item{x}{string Column name of data to place on x-axis}

\item{ymin}{string Column name of lower whisker extent for each box}

\item{lower}{string Column name of lower border value of each box}

\item{upper}{string Column name of upper border value of each box}

\item{ymax}{string Column name of upper whisker extent for each box}

\item{median}{string Column name of median value of each box}

\item{fill}{string (optional) Column name of fill value of each box}

\item{color}{string (optional) Column name of color value of each box (outline)}

\item{x_group}{string (optional) Column name of column to concatenate
with x-column to make each x value
unique (e.g. duration)}

\item{ann}{string (optional, default = NULL) Column name of text annotation
for each box}

\item{ann_position_adjust}{(optional) default = 0.4 The y-displacement of
annotations, if ann is specified. Increase value
to move annotations further from box.}

\item{x_lab}{string (optional) X-axis label}

\item{y_lab}{string (optional) Y-axis label}

\item{fill_lab}{string (optional) Legend label of fill category}

\item{color_lab}{string (optional) Legend label of color category}

\item{title}{string (optional) Title of plot}

\item{hline_int}{int (optional) X intercept of hline if one is desired. If
not specified, no hline will be drawn.}

\item{coord_flip}{(optional) If TRUE (defualt), coordinates are flipped (boxes are
oriented horizontally).}
}
\description{
Use alongside boxplotModUI() to add modified boxplot to app.
Optional features include annotations, and an hline.
}
\details{
Modified boxplot allows user to specify extent of box limits (lower and upper)
and whisker limits (ymin and ymax)vby passing columns names with minimum and
maximum extent values. Data should be reactive, and include columns with "upper",
"lower", "ymin", "ymax" and "median" values already calculated.
E.g. "upper" might be the 90th percentile for an x value and "lower"
could be the 10th percentile.  If specified, the plot's coordinates may be
flipped (e.g. \code{x} and \code{x_lab} will appear on the y-axis, and boxes will be horizontally
oriented.)
}
