% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotModUI.R
\name{boxplotModUI}
\alias{boxplotModUI}
\title{Boxplot Mod UI Function}
\usage{
boxplotModUI(id)
}
\arguments{
\item{id}{string ID label that links boxplotModUI() to boxplotModServer()}
}
\value{

}
\description{
Use alongside boxplotModServer() to add modified boxplot to app.
Modified boxplot allows user to specify extent of box limits and whisker
limits by passing columns with minimum and maximum extent values.
}
