% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotServer.R
\name{boxplotServer}
\alias{boxplotServer}
\title{Boxplot Server Function}
\usage{
boxplotServer(
  id,
  data,
  x,
  y,
  fill = NULL,
  ann = NULL,
  agg_ann = NULL,
  agg_fun = NULL,
  x_lab = "",
  y_lab = "",
  fill_lab = "",
  title = "",
  hline_int = NA_real_,
  hline_ann = NA,
  coord_flip = FALSE,
  position_dodge = 0.5,
  position_dodge_ann = 0.75,
  ann_vjust = 1,
  plot_specs = NULL
)
}
\arguments{
\item{id}{string ID label that links boxplotUI() to boxplotServer()}

\item{data}{reactive Reactive dataframe}

\item{x}{string Column name of data categories to place on x-axis}

\item{y}{string Column name of values used to calculate boxes per x value}

\item{fill}{string (optional) Column name of fill value of each box}

\item{ann}{string (optional, default = NULL) Column name of text annotation
for each box}

\item{agg_ann}{string (optional) Column name with values to aggregate and
use as labels per each x value}

\item{agg_fun}{FUN (optional) Function used to aggregate agg_ann when
calculating labels per box (e.g. sum, min, max)}

\item{x_lab}{string (optional) X-axis label}

\item{y_lab}{string (optional) Y-axis label}

\item{fill_lab}{string (optional) Legend label of fill category}

\item{title}{string (optional) Title of plot}

\item{hline_int}{int (optional) X intercept of hline if one is desired. If
not specified, no hline will be drawn.}

\item{hline_ann}{string (optional) Label for hline}

\item{coord_flip}{(optional) If TRUE (defualt), coordinates are flipped
(boxes are oriented horizontally).}

\item{position_dodge}{numeric (optional) The offset between boxes plotted for
the same x-value. Default is 0.5.}

\item{position_dodge_ann}{numeric (optional) The offset between labels
plotted for the same x-value. Default=0.75.}

\item{ann_vjust}{numeric (optional) The vertical adjustment of optional box
labels. Default is 1.}

\item{plot_specs}{(optional) obj Additional GGplot elements to add to plot
e.g. theme(axis.text.x = element_text(angle = 90)).
If multiple elements are to be added, must be
a vector (e.g. c(ylim(0,4),
hline(y_intercept = 0, color = 'red))).}
}
\value{

}
\description{
Use alongside boxplotUI() to add boxplot to app.
Upper box boundary is 75th percentile, lower boundary is 25th percentile.
Whiskers extend to +/- 1.5*IQR beyond the upper/lower box boundaries in either
direction.
}
\details{
Optional features include annotations, and an hline.
Data should be reactive, If specified, the plot's coordinates may be
flipped (e.g. \code{x} and \code{x_lab} will appear on the y-axis, and boxes will be
horizontally oriented.)
}
