% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotUI.R
\name{boxplotUI}
\alias{boxplotUI}
\title{Boxplot UI Function}
\usage{
boxplotUI(id)
}
\arguments{
\item{id}{string ID label that links boxplotUI() to boxplotServer()}
}
\value{

}
\description{
Use alongside boxplotServer() to add boxplot to app.
}
