% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mapping.R
\name{color_mapping}
\alias{color_mapping}
\title{Create color map}
\usage{
color_mapping(num_col, exclude_achromatic = FALSE)
}
\arguments{
\item{num_col}{integer, number of unique values to assign a color to}

\item{exclude_achromatic}{boolean, exclude black and gray colors?}
}
\description{
Create color map
}
