% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_template.R
\name{dashboard_template}
\alias{dashboard_template}
\title{Creates template for STi dashboard with tab panel}
\usage{
dashboard_template(
  filename,
  tab_num,
  tab_names,
  tab_groups = NULL,
  app_title = NULL,
  header_img = NULL,
  readme_file = "",
  PAL = "sonoma_color_palette",
  CB_PAL = "viridis::magma",
  PAL_a = 1
)
}
\arguments{
\item{filename}{string Name of file (should follow format app_*.R). Must not
already exist.}

\item{tab_num}{integer Number of tabs}

\item{tab_names}{character vector Names of tabs}

\item{tab_groups}{integer vector (optional) Desired tab groupings. Each group will have
a unique color. Must sum to tab_num. e.g.
If tab_num = 5, valid tab_groups
include c(2,3), c(1,4), c(3,2), etc. If no
tab groups are specified (default), all
tabs will be the same color.}

\item{app_title}{string (optional) Title of App to be written in header. If not specified, title is blank.}

\item{header_img}{string (optional) Filename of image to include in title header. Must
be in directory named www.If not specified, no image is added.}

\item{readme_file}{string (optional) Filename of README file. Must be in the www folder.}

\item{PAL}{string (optional) The name of the non-colorblind-palette function
to use (default = sonoma_color_palette)}

\item{CB_PAL}{string (optional) The name of the colorblind-palette function
to use (default = sonoma_color_palette)}

\item{PAL_a}{double (optional) An alpha value between 0 and 1 to set opacity (default = 1)}
}
\value{

}
\description{
Code template will be written according to specifications below into
a file and opened in RStudio.
}
