% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionLimitsPlotServer.R
\name{detectionLimitsPlotServer}
\alias{detectionLimitsPlotServer}
\title{Detection Limits Plot Server}
\usage{
detectionLimitsPlotServer(
  id,
  input_data,
  x,
  y,
  label,
  color,
  size,
  x_lab = "Percent above detection limit (\%)",
  y_lab = "Detection limit (median in units of measure)",
  color_lab = NULL,
  size_lab = NULL,
  title = "",
  filter_col = NULL,
  filter_rct = shiny::reactive({     NULL })
)
}
\arguments{
\item{id}{string ID label that links detectionLimitsPlotUI() to detectionLimitsPlotServer()}

\item{input_data}{reactive Reactive dataframe with columns that determine x
and y position, label, size, and color.}

\item{x}{string Column name of x-position}

\item{y}{string Column name of y-position}

\item{label}{string Column name with labels to be plotted}

\item{color}{string Column name that determines color of plotted label}

\item{size}{string Column name that determines size of plotted label (numeric)}

\item{x_lab}{string (optional) X-axis label, default = 'Percent above detection limit (\%)'}

\item{y_lab}{string (optional) Y-axis label, default = 'Detection limit (median in units of measure)'}

\item{color_lab}{string (optional) Title for color category in legend}

\item{size_lab}{string (optional) Title for size category in legend}

\item{title}{string (optional) Plot title}

\item{filter_col}{string (optional) Column used to filter data based on
specified \code{filter_rct}.}

\item{filter_rct}{reactive Value in \code{filter_col} to filter dataframe (equal).}
}
\value{
character vector Method codes displayed in the plot (updated by brush)
}
\description{
Used alongside detectionLimitsPlotUI() to add a detection limits plot to app.
Text labels are plotted in a position determined by the x and y arguments,
colored by the color argument, and sized by the size argument.
}
