% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionLimitsPlotUI.R
\name{detectionLimitsPlotUI}
\alias{detectionLimitsPlotUI}
\title{Detection Limits Plot UI}
\usage{
detectionLimitsPlotUI(id)
}
\arguments{
\item{id}{string ID label that links detectionLimitsPlotUI() to detectionLimitsPlotServer()}
}
\value{

}
\description{
Used alongside detectionLimitsPlotServer() to add a detection limits plot to app.
Method labels are plotted according to MDL and \% of observations above the
MDL, colored by parameter and sized by number of observations.
}
