% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footerDateofDataUpdateUI.R
\name{footerDateofDataUpdateUI}
\alias{footerDateofDataUpdateUI}
\title{Date of Last Update Footer}
\usage{
footerDateofDataUpdateUI(dt_df)
}
\arguments{
\item{dt_df}{data.frame Must contain two columns, \code{weekly} and \code{quarterly},
with the most recent date of data update per
download cadence in the first row.}
}
\description{
Paste at bottom of UI to add a note listing the most recent deployment date
}
