% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geompatternModServer.R
\name{geompatternModServer}
\alias{geompatternModServer}
\title{Boxplot/Barplot with pattern server function}
\usage{
geompatternModServer(
  id,
  data,
  x,
  y,
  fill_col,
  legend_title = "",
  geom_plot_type = c("bar", "box"),
  xlab = "",
  ylab = "",
  title = "",
  caption = "",
  subtitle = "",
  pattern_map,
  color_map,
  breaks,
  second_axis = FALSE,
  coord_flip = FALSE,
  bar_stack = FALSE,
  sort_x_by,
  ...
)
}
\arguments{
\item{id}{id that links server to UI function}

\item{data}{reactive data.frame}

\item{x}{x-axis column name as a string}

\item{y}{y-axis column name as a string}

\item{fill_col}{fill column name as a string}

\item{legend_title}{title of legend}

\item{geom_plot_type}{either "bar" or "box"}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{plot title}

\item{caption}{plot caption (optional)}

\item{subtitle}{plot subtitle (optional)}

\item{pattern_map}{pattern map (from pattern_mapping() function)}

\item{color_map}{color map (from color_mapping() function)}

\item{breaks}{vector of breaks to map onto map pattern and colors map vectors (should be same length as pattern_map and color_map)}

\item{second_axis}{boolean, display second identical y-axis?}

\item{coord_flip}{boolean, flip figure?}

\item{bar_stack}{boolean, should the bar plot be stacked or dodged?}

\item{sort_x_by}{optional vector to sort the x-axis (should be the same length as x-axis column)}

\item{...}{any argument passed to geom_bar_pattern or geom_boxplot_pattern}
}
\description{
Creates a barplot or boxplot with pattern fill depending on number of variables
}
