% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geompatternModUI.R
\name{geompatternModUI}
\alias{geompatternModUI}
\title{Boxplot/Barplot with pattern UI function}
\usage{
geompatternModUI(id, ...)
}
\arguments{
\item{id}{id that links server function to ui function}

\item{...}{other arguments to plotOutput (e.g. height, width)}
}
\description{
Creates a barplot or boxplot with pattern fill depending on number of variables
}
