% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplotServer.R
\name{lineplotServer}
\alias{lineplotServer}
\title{Lineplot Server Function}
\usage{
lineplotServer(
  id,
  input_data,
  point = FALSE,
  x,
  y,
  units,
  color,
  line_width = 1,
  point_size = 2,
  linetype = NULL,
  label = NULL,
  label_size = 8,
  label_vjust = -1.05,
  title = "",
  x_lab = "",
  y_lab = "",
  color_lab = "",
  linetype_lab = "",
  filter_rct = shiny::reactive({     NULL }),
  filter2_rct = shiny::reactive({     NULL }),
  filter_col = NULL,
  filter2_col = NULL,
  highlight = FALSE,
  highlight_col = "AQS_SITECODE",
  highlight_value_rct = shiny::reactive({     NULL }),
  x_numeric = TRUE,
  corner_ann = "Select site on map to highlight",
  plot_specs = NULL
)
}
\arguments{
\item{id}{string ID label that links lineplotUI() to lineplotServer()}

\item{input_data}{reactive Reactive dataframe}

\item{point}{boolean (optional) If TRUE (default), points will be plotted over
lines}

\item{x}{string Column with values to plot on x-axis}

\item{y}{string Name of values to plot on y-axis}

\item{units}{string Name of column with units per value}

\item{color}{string Name of column with groups to separate data into
distinct lines (e.g. one line per sitecode)}

\item{line_width}{numeric (optional) Width of lines}

\item{point_size}{numeric (optional) Size of points, if points = TRUE}

\item{linetype}{string (optional) Name of column that determines linetype. If null,
(default) all column will be solid.}

\item{label}{string (optional) Name of column with labels to print above every
data point}

\item{label_size}{integer (optional) Size of label text, if label is specified. (default = 8)}

\item{label_vjust}{numeric (optional) Vertical adjustment of labels compared
to position of observation, if label is specified. (default = -1.05)}

\item{title}{string (optional) Plot title}

\item{x_lab}{string (optional) X-axis label. If not specified, left blank}

\item{y_lab}{string (optional) Y-axis label.
If not specified, defaults to the first value
in the filtered data's units column.}

\item{color_lab}{string (optional) Legend label for color categories}

\item{linetype_lab}{string (optional) Legend label for linetype categories}

\item{filter_rct}{reactive Parameter to plot}

\item{filter2_rct}{reactive (optional) Duration to plot. If spcified, input_data
will be filtered to the passed duration.}

\item{filter_col}{string Column name of parameter names}

\item{filter2_col}{string (optional) Column name of durations}

\item{highlight}{boolean (optional) If TRUE, all lines but highlight line will be
gray. If false, colors applied as normal
aes.}

\item{highlight_col}{string Column name with value to determine highlighted line(s)}

\item{highlight_value_rct}{reactive (optional) Value to highlight (equality)}

\item{x_numeric}{boolean (optional, default = TRUE) If TRUE, x-axis is
specified as continuous and pretty_breaks() is called.}

\item{corner_ann}{string (optional) Annotation to put in the top-right corner
when no highlight value is specified. Default is
"Select site on map to highlight"}

\item{plot_specs}{(optional) obj Additional GGplot elements to add to plot
e.g. theme(axis.text.x = element_text(angle = 90)).
Can be string of added elements.}
}
\value{

}
\description{
Use alongside lineplotUI() to add lineplot of concentration data to app.
Specified color determines how to separate data into distinct lines (
e.g. one line per AQS_SITECODE). Specify a reactive parameter and duration to filter
\code{input_data}. Specify \code{highlight = TRUE} and a reactive \code{highlight_value_rct}, a value from the
\code{highlight_col} column, to highlight the line for that group. Overlay points
over lines using the \code{point} argument.
}
