% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplotUI.R
\name{lineplotUI}
\alias{lineplotUI}
\title{Lineplot Server Function}
\usage{
lineplotUI(id)
}
\arguments{
\item{id}{string ID label that links lineplotUI() to lineplotServer()}
}
\description{
Use alongside lineplotServer() to add lineplot of concentration data to app.
}
