% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSelectHeader.R
\name{mapSelectHeader}
\alias{mapSelectHeader}
\title{Header table for selected map site}
\usage{
mapSelectHeader(id, data_row, select_cols, col_labels)
}
\arguments{
\item{id}{id used to link UI function to Server}

\item{data_row}{reactive Row of data to be displayed, e.g. the $sitedata
output of sitemapServer()}

\item{select_cols}{character vector Columns to display in table}

\item{col_labels}{character vector Column labels to display in table}
}
\value{
function Call to DT::renderDataTable
}
\description{
Displays sspecified columns from data passed to sitemapServer function
for site clicked on map. Must be set to output element with a unique ID,
with a call to mapSelectHeaderUI specifying the same ID.
e.g. output$header <- mapSelectHeader()
}
