% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCompletenessServer.R
\name{networkCompletenessServer}
\alias{networkCompletenessServer}
\title{Network Completeness Plot Server}
\usage{
networkCompletenessServer(
  id,
  input_data,
  x,
  segment = TRUE,
  seg_min = NULL,
  seg_max = NULL,
  y,
  color = NULL,
  param_col = NULL,
  param_code_col = NULL,
  param_code_list = NULL,
  table_cols = NULL,
  table_col_labels = NULL,
  xlab = "Rank ordered site completeness",
  ylab = "Completeness (\%)",
  title_rct = reactive({     NULL }),
  hline_int = NA_real_,
  hline_ann = "",
  hline_color = "red",
  filter_col = NULL,
  filter_rct = shiny::reactive({     NULL }),
  highlight_col = NULL,
  highlight_rct = shiny::reactive({     NULL }),
  dragbox_x_only = F,
  corner_ann = "",
  shape_values = NULL,
  color_values = NULL,
  plot_specs = NULL,
  subset_graph_col = NULL
)
}
\arguments{
\item{id}{string ID that links networkCompletenessUI() to networkCompletenessServer()}

\item{input_data}{reactive Reactive data frame}

\item{x}{string Column name of x-axis value}

\item{segment}{boolean (optional) If TRUE (default), a vertical segment will
be drawn. Requires seg_min, seg_max values to be filled.}

\item{seg_min}{string (optional) Required if segment is TRUE.
Column name of segment minimum value}

\item{seg_max}{string (optional) Required if segment is TRUE.
Column name of segment maximum value}

\item{y}{string Column name of y-axis value of points}

\item{color}{string (optional) Column name of value to color points by}

\item{param_col}{string (optional) Column name for parameter name}

\item{param_code_col}{string (optional) Column name for parameter code.
Should be specified if passing a subset of
parameter codes in \code{param_code_list}.}

\item{param_code_list}{character vector (optional) List of parameter codes to plot.}

\item{table_cols}{character vector (optional) Columns to include in table
that appears when plot is brushed. If not
specified, all columns are included.}

\item{table_col_labels}{character vector (optional) Specific column labels to
print in table. Must match in order
with 'table_cols'. If not specified,
column names in input_data will
be used.}

\item{xlab}{string (optional) Label for x-axis. Default = 'Rank ordered site completeness'.}

\item{ylab}{string (optional) Label for y-axis. Default = 'Completeness (\%)'.}

\item{title_rct}{reactive (optional) Plot title. Must be reactive. Default = reactive({NULL}).}

\item{hline_int}{integer (optional) Y-intercept of horizontal line. Default = NA.}

\item{hline_ann}{string (optional) Annotation for optional horizontal line. Default = ''.}

\item{hline_color}{string (optional) Color of hline (if hline int specified)}

\item{filter_col}{string (optional) Column name of column to filter data by. Default = NULL.}

\item{filter_rct}{reactive (optional) Value to filter data by. Default = reactive({NULL}).}

\item{highlight_col}{string (optional) Column name of column that determines segment to highlight. Default = NULL.}

\item{highlight_rct}{reactive (optional) Value to highlight. Default = reactive({NULL}).}

\item{dragbox_x_only}{boolean (optional) Default = F. If TRUE, table will be filtered
only by the x-range of the dragbox (e.g.
all y values in the selected x-range will
be shown in the table).}

\item{corner_ann}{string (optional) Annotation in plot corner when no value is highlighted.}

\item{shape_values}{vector (optional) A vector of the full set of parameters
to assign a standard shape to. Useful
if input_data is reactively filtered
and consistent shapes are desired.
Must contain all possible values for
parameter that may appear in the plot.}

\item{color_values}{vector (optional) A vector of the full set of values
in the specified color column
to assign a standard color to. Useful
if input_data is reactively filtered
and consistent colors are desired.
Must contain all possible values for
color that may appear in the plot.}

\item{plot_specs}{object (optional) Additional ggplot elements to add to the
plot (e.g. ylim(0,100)). Can be multiple elements
added together.}

\item{subset_graph_col}{string (optional) If specified, input_data must contain
a boolean column that indicates a subset of rows
that may appear on the graph. Only data with a
value of TRUE in this column will be graphed,
though all data values will appear in table that meet
the brush specifications. e.g. Can be used to plot
only the lowest POC at a site, while showing
information on all POCs in the details table.}
}
\value{

}
\description{
Used alongside networkCompletenessUI() to add a network completeness
plot to app. For each x-value (site), a vertical segment from seg_min to seg_max is
plotted, typically representing max and min completeness values across
parameters for a site. Overlaid on these segments are points representing
completeness value for each parameter measured at a site. A certain x-value
can be highlighted with an arrow and label. A horizontal line may be added
at a given y-intercept. A reactive filter column and value may be passed to
filter plotted data.
}
