% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCompletenessUI.R
\name{networkCompletenessUI}
\alias{networkCompletenessUI}
\title{Network Completeness Plot UI}
\usage{
networkCompletenessUI(id, add_brush = TRUE)
}
\arguments{
\item{id}{string ID that links networkCompletenessUI() to networkCompletenessServer()}

\item{add_brush}{boolean If TRUE (default), brush and double-click added to
rendered plot. This allows users to display details
of selected data below the plot.}
}
\description{
Used alongside networkCompletenessUI() to add a network completeness
plot to app.
}
