% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_to_expression.R
\name{param_to_expression}
\alias{param_to_expression}
\title{Convert units to formatted expression}
\usage{
param_to_expression(parameter)
}
\arguments{
\item{parameter}{character vector A string or vector of parameter names to be
converted}
}
\value{
character vector Formatted expressions
}
\description{
Convert units to formatted expression
}
