% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_mapping.R
\name{pattern_mapping}
\alias{pattern_mapping}
\title{Create pattern map}
\usage{
pattern_mapping(num_col, exclude_achromatic = FALSE)
}
\arguments{
\item{num_col}{integer, number of unique values to assign a pattern to}

\item{exclude_achromatic}{boolean, exclude black and gray colors?}
}
\description{
Create pattern map
}
