% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutionRoseTabUI.R
\name{pollutionRoseTabUI}
\alias{pollutionRoseTabUI}
\title{Pollution Rose Tab UI}
\usage{
pollutionRoseTabUI(
  id,
  season_filter = TRUE,
  year_filter = TRUE,
  dow_filter = FALSE,
  hour_filter = FALSE,
  metric_filter = TRUE,
  num_plots = 2
)
}
\arguments{
\item{id}{string ID used to link pollutionRoseTabUI() to pollutionRoseTabServer()}

\item{season_filter}{boolean If TRUE, multi-select input for season is added.
Seasons include spring, summer, fall, winter.}

\item{year_filter}{boolean If TRUE, multi-select input for year is added.
Years range from 2019-2020.}

\item{dow_filter}{boolean If TRUE, multi-select input for day of week is added.}

\item{hour_filter}{boolean If TRUE, multi-select input for hour is added.}

\item{metric_filter}{boolean (optional) If TRUE (default) filter
is added on tab to display either
concentrations or AQI index.}

\item{num_plots}{integer Number of plots to show on the PR tab. Maximum of 4.}
}
\value{

}
\description{
Used alongside pollutionRoseTabUI() to add a collection of reactive
elements to an app, including multiselects for seasons and years, a select
for metric type (aqi or concentration), and up to 2 pollution roses for
two specified parameters.
}
