% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutionRoseUI.R
\name{pollutionRoseUI}
\alias{pollutionRoseUI}
\title{Pollution Rose UI}
\usage{
pollutionRoseUI(id)
}
\arguments{
\item{id}{string ID used to link pollutionRoseUI() to pollutionRoseServer()}
}
\description{
Used alongside pollutionRoseServer() to add a pollution rose with 1hr data to the app.
}
