% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_sonoma.R
\name{scale_color_sonoma}
\alias{scale_color_sonoma}
\title{STi ggplot discrete color scheme}
\usage{
scale_color_sonoma(reverse = FALSE)
}
\arguments{
\item{reverse}{boolean If TRUE, reverses order of color palette}
}
\value{

}
\description{
Color palette defined in sonoma_color_palettea(). Applies to 'color' aesthetic
in ggplot object.
}
