% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_sonoma.R
\name{scale_fill_sonoma}
\alias{scale_fill_sonoma}
\title{STi ggplot discrete fill scheme}
\usage{
scale_fill_sonoma(reverse = FALSE)
}
\arguments{
\item{reverse}{boolean If TRUE, reverses order of color palette}
}
\description{
Color palette defined in sonoma_color_palette(). Applies to 'color' aesthetic
in ggplot object.
}
