% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterServer.R
\name{scatterServer}
\alias{scatterServer}
\title{Scatter plot Server}
\usage{
scatterServer(
  id,
  data_param1_rct,
  data_param2_rct,
  param_col,
  measurement_col,
  key_col,
  unit_col,
  poc_col,
  param_rct,
  sec_param_rct,
  brush = shiny::reactive({     NULL }),
  legend_pos = c(0, 1)
)
}
\arguments{
\item{id}{string ID that links scatterUI() to scatterServer()}

\item{data_param1_rct}{reactive Reactive dataframe with columns for parameter,
measurement and unit with a key used to join
data1 to data2. Should only contain data for
param_1_rct.
spreading the dataframe by parameter.}

\item{data_param2_rct}{reactive Reactive dataframe with columns for parameter,
measurement and unit with a key used to join
data1 to data2. Should only contain data for
param_2_rct.}

\item{param_col}{string Column name of column with parameter values}

\item{measurement_col}{string Column name of column with measurement values}

\item{key_col}{string Column name of column with key values used in spread
operation, and column affected by
brush (e.g. datetime).}

\item{unit_col}{string Column name of column with unit values}

\item{poc_col}{string Column name of column iwth poc number}

\item{param_rct}{reactive Parameter value to plot on x-axis}

\item{sec_param_rct}{reactive Parameter value to plot on y-axis}

\item{brush}{reactive (optional) Brush object used to adjust displayed data.
Must have elements named xmin, xmax,
ymin, ymax.}

\item{legend_pos}{numeric vector (optional) Default = c(1,0), offset from
upper-left corner that legend should be
positioned in.}
}
\value{

}
\description{
Used alongside scatterUI() to add a scatter plot to app. Takes in a
long-formatted dataframe with a column for Parameter, Unit, and
Measurement. Spreads the dataframe by the parameter column
using \code{key_col} as the key (typically a timestamp).
Optionally accepts a brush (e.g. from a timeseries plot) and filters
the \code{key_col} accordingly. Optionally plots a 1-to-1 a-b line and a
linear regression line.
}
