% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterUI.R
\name{scatterUI}
\alias{scatterUI}
\title{Scatter plot UI}
\usage{
scatterUI(id, add_regression_option = TRUE, add_one2one_option = TRUE)
}
\arguments{
\item{id}{string ID that links scatterUI() to scatterServer()}

\item{add_regression_option}{bool (optional) If TRUE (default),
adds checkbox to add linear regression line
to scatterplot.}

\item{add_one2one_option}{bool (optional) If TRUE (default),
adds checkbox to add one-to-one line
to scatterplot.}
}
\description{
Used alongside scatterServer() to add a scatter plot to app.
}
