% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitemapServer.R
\name{sitemapServer}
\alias{sitemapServer}
\title{Map Server Function}
\usage{
sitemapServer(
  id,
  data,
  filter_col = NULL,
  color_col = NULL,
  color_order = NULL,
  label_col,
  cb = reactive({     NULL }),
  click_buffer = 0.003,
  add_polygon = FALSE,
  polygon_df = reactive({     NULL }),
  polygon_label_col = NULL,
  polygon_color = "navy",
  polygon_weight = 3
)
}
\arguments{
\item{id}{string ID label that links sitemapUI() to sitemapServer()}

\item{data}{reactive data.frame Reactive data with location info to be plotted on map. Must
contain columns lat, lng, and AQS_SITECODE corresponding
to lattude, longitude, and aqs site code respectively.}

\item{filter_col}{string (optional, default = NULL) Column to filter sites
shown on map by. Should only have 2 values. If specified,
fluidRow with radiobuttons labeled by \code{filter_labels}
will be added above map. If specified, \code{filter_labels}
must also be specified.}

\item{color_col}{string (optional, default = NULL) Column used to color site
markers on map.}

\item{color_order}{character vector (optional) A vector with the order of
color values that appear in the legend.
All values in color_col must be specified.
1st specified will be blue, then green,
the yellow, etc (according to sonoma_
color_palette).}

\item{label_col}{string Column with labels for each circle}

\item{cb}{reactive (optional) If not NULL, colorblind palette will be applied}

\item{click_buffer}{numeric (optional) Buffer around a click to select the
closest map marker. Default = 0.003 degrees.}

\item{add_polygon}{boolean (optional) Defalt = FALSE. Set to TRUE if a polygon
will be added to the map in addition to cirle_markers. If TRUE,
polygon_df and polygon_label_col must be specified.}

\item{polygon_df}{reactive SpatialPolygonsDataFrame A reactive spatial dataframe,
must be specified if
add_polygon is TRUE. A SpatialPolygonsDataFrame, e.g. as
imported by rgdal's readOGR function, with polygon geometries
and an attribute for polygon_label_col.}

\item{polygon_label_col}{string (optional) Must be specified if add_polygon
is TRUE. A string specifying the attribute in
polygon_df to label polygons by on the map.}

\item{polygon_color}{string (optional) Default = 'navy'. Color for the
polygon outline.}

\item{polygon_weight}{integer (optional) Weight of the polygon outline.
Default = 3.}
}
\value{
list List with selected filter value ($radio), and selected label ($sitecode)
}
\description{
Use alongside sitemapUI() to add map of aqs sites to app. Map will include
tooltip with aqs site code, and optionally a
radiobutton filter of sites and/or legend with site categories.
}
