% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitemapUI.R
\name{sitemapUI}
\alias{sitemapUI}
\title{Leaflet Map UI Function}
\usage{
sitemapUI(
  id,
  filter_col = NULL,
  filter_labels = NULL,
  top_message = "Click a circle on the map to select a site"
)
}
\arguments{
\item{id}{string ID label that links sitemapUI() to sitemapServer()}

\item{filter_col}{string (optional, default = NULL) Column to filter sites
shown on map by. Should only have 2 values. If specified,
fluidRow with radiobuttons labeled by \code{filter_labels}
will be added above map. If specified, \code{filter_labels}
must also be specified.}

\item{filter_labels}{character vector (optional, default = NULL) Must be
specified if \code{filter_col} is filled. Labels applied to
radio buttons that specify sites to be displayed
on map. Initial value is 1st label in vector.
Should be length 2.}

\item{top_message}{string Text to display above map. Default message is
"Click a circle on the map to select a site". If no
message desired, input an empty string.}
}
\value{

}
\description{
Used alongside sitemapServer() to add a map of aqs sites (wrapped in a fluidRow)
to an app. Map will include tooltip with aqs site code, and optionally a
radiobutton filter of sites and/or legend with site categories.
}
