% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonoma_table_cb_palette.R
\name{sonoma_table_cb_palette}
\alias{sonoma_table_cb_palette}
\title{STi graphics colorblind palette for tables}
\usage{
sonoma_table_cb_palette(c, a = 1, show = FALSE)
}
\arguments{
\item{c}{integer (optional) An integer (1-9) specifying
the number of colors in the palette to return.}

\item{a}{double (optional) An alpha value (0-1) to adjust the opacity of the
color palette}

\item{show}{boolean (optional, defalut = FALSE) If TRUE, outputs color
palette in plots tab}
}
\value{
string list
}
\description{
STi graphics colorblind palette for tables
}
\examples{
sonoma_table_cb_palette(3)
}
