% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_color_palette.R
\name{tab_color_palette}
\alias{tab_color_palette}
\title{STi tab color palette}
\usage{
tab_color_palette(n, a = 1, cb = F, font = F, show = F)
}
\arguments{
\item{n}{integer (optional) An integer (1-3) specifying
the number of colors in the palette to return.}

\item{a}{double (optional) An alpha value (0-1) to adjust the opacity of the
color palette}

\item{cb}{boolean (optional) Default = False, whether or not the palette
should be colorblind friendly.}

\item{font}{boolean (optional) Default = False, whether or not to return the
font color for the relevant tab colors (ordered).}

\item{show}{boolean (optional, defalut = FALSE) If TRUE, outputs color
palette in plots tab}
}
\value{
string list
}
\description{
STi tab color palette
}
\examples{
sonoma_color_palette(3)
}
