% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabColorKeyUI.R
\name{tabColorKeyUI}
\alias{tabColorKeyUI}
\title{Tab color key UI}
\usage{
tabColorKeyUI(id)
}
\arguments{
\item{id}{string Links call to tabColorKeyUI to tabColorKeyServer function}
}
\description{
Creates key for tab colors that updates depending on coloblind accessible
selection.
}
