% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableServer.R
\name{tableServer}
\alias{tableServer}
\title{Data Table Server Function}
\usage{
tableServer(
  id,
  input_data,
  page_length = 5,
  select_cols = NULL,
  col_label_names = NULL,
  arrange_col = NULL,
  filter_col = NULL,
  filter_rct = shiny::reactive({
     list(NULL)
 }),
  color_col = NULL,
  color_order = NULL,
  caption_rct = reactive({
     NULL
 }),
  pdf_button = TRUE,
  cb = shiny::reactive({
     NULL
 }),
  load_all_data = TRUE,
  download_filename = "data",
  dom_spec = "Bfrtip"
)
}
\arguments{
\item{id}{string ID label that links tablueUI() to tableServer()}

\item{input_data}{reactive Reactive dataframe}

\item{page_length}{integer (optional, default = 5) Number of items to display
at once in table (if < nrow(input_data), table
is paginated)}

\item{select_cols}{character (optional, default = NULL) Vector of columns
to including in data table.}

\item{col_label_names}{character (optional, default = NULL) Vector of column
header names to appear in data frame.
Must match number of columns in input_data
or `select_cols`` (if specified).}

\item{arrange_col}{string (optional, default = 1st column) Column to arrange
table by.}

\item{filter_col}{string (optional, default = NULL) Name of column to apply
reactive filter to...reactive filter value passed
in \code{filter_rct}. If specified, filter_rct must
be supplied. If not specified, table has no
reactive filtering.}

\item{filter_rct}{reactive list (optional, default = NULL) Reactive list of values
to filter \code{filter_col} by. If specified,
filter_col must be supplied.If not specified,
table has no reactive filtering.}

\item{color_col}{string (optional, default = NULL) Column that determines row
colors. If not specified, columns will be alternating
grays.}

\item{color_order}{character vector (optional) A vector with the order of
color values that appear in the legend.
All values in color_col must be specified.
1st specified will be blue, then green,
the yellow, etc (according to sonoma_
color_palette).}

\item{caption_rct}{reactive string (optional, default = NULL) Caption for data table.}

\item{pdf_button}{boolean (optional, default = TRUE) If true, "download
table" button added.}

\item{cb}{reactive (optional) If not NULL, colorblind palette will be applied}

\item{load_all_data}{boolean (optional) If TRUE (default), all data in table
will be loaded and available for download. If false,
only first page (visible) data is loaded and available
for download. Not recommended for large datasets.}

\item{download_filename}{string (optional) Defaults to 'data', the filename
desired for downloaded table data.}

\item{dom_spec}{sting (optional) Default = 'Bfrtip', establishes built-in
table control elements. https://datatables.net/reference/option/dom}
}
\description{
Data Table Server Function
}
