% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableUI.R
\name{tableUI}
\alias{tableUI}
\title{Data Table UI funtion}
\usage{
tableUI(id)
}
\arguments{
\item{id}{string ID label that links tablueUI() to tableServer()}
}
\value{

}
\description{
Used alongside tableServer() to add a data table (wrapped in a fluidRow)
to an app.
}
