% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sonoma.R
\name{theme_sonoma}
\alias{theme_sonoma}
\title{STi ggplot theme}
\usage{
theme_sonoma()
}
\value{

}
\description{
Centered & bold title, 14 base font size, white background,
major and minor gridlines, legend on right, black plot border
}
