% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesInvestigationTabServer.R
\name{timeseriesInvestigationTabServer}
\alias{timeseriesInvestigationTabServer}
\title{Timeseries Investigation Tab Server}
\usage{
timeseriesInvestigationTabServer(
  id,
  data_1hr_rct,
  data_2_rct,
  data_1hr_wind_rct,
  data_24hr_wind_rct,
  selected_site_rct,
  paramlist_1hr_rct,
  paramlist_2_rct,
  dt_col,
  param_col,
  unit_col,
  measurement_col,
  sampledur_col,
  poc_col,
  wd_col,
  ws_col,
  site_col,
  linesize = 0.75,
  pointsize = 2
)
}
\arguments{
\item{id}{string ID the connects timeseriesInvestigationTabUI() to timeseriesInvestigationTabServer()}

\item{data_1hr_rct}{reactive data.frame 1-hr concentration data with columns
for datetime, parameter name, units,
concentration measurements, sitecode.}

\item{data_2_rct}{reactive data.frame Concentration data for non-1hr data with columns
for datetime, parameter name, units,
concentration measurements, sample duration,
sitecode.}

\item{data_1hr_wind_rct}{reactive data.frame 1hr Wind data with columns
for datetime, sitecode, wind direction,
wind speed.}

\item{data_24hr_wind_rct}{reactive data.frame 24hr Wind data with columns
for datetime, sitecode, wind direction,
wind speed.}

\item{selected_site_rct}{reactive string Selected AQS Sitecode}

\item{paramlist_1hr_rct}{reactive vector List of 1hr parameters for selected site}

\item{paramlist_2_rct}{reactive vector List of non- 1-hr parameters for selected site.
Must have column for sample_duration.}

\item{dt_col}{string Name of datetime column in concentration data}

\item{param_col}{string Name of column with parameter name values in concentration AND wind data}

\item{unit_col}{string Name of column with unit values in concentration data}

\item{measurement_col}{string Name of concentration column for timeseries}

\item{sampledur_col}{string Name of sample duration column in data_2_rct}

\item{poc_col}{string Column name of column with poc number.}

\item{wd_col}{string Name of column in wind data with direction values}

\item{ws_col}{string Name of column in wind data with speed values}

\item{site_col}{string Name of column in wind AND concentration data with values
of AQS sitecode.}

\item{linesize}{numeric (optional) Size of plotted line (default = 0.5) on TS}

\item{pointsize}{numeric (optional) Size of plotted points (default = 3) on TS}
}
\description{
Used alongside timeseriesInvestigationTabUI() to add a collection of
reactive elements to the app, including multiple parameter selections,
a timeseries plot, a scatter plot, and a pollution rose. Wrapper for standalone
functions scatterServer(), timeseriesServer() and pollutionRoseServer().
}
