% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesInvestigationTabUI.R
\name{timeseriesInvestigationTabUI}
\alias{timeseriesInvestigationTabUI}
\title{Timeseries Investigation Tab UI}
\usage{
timeseriesInvestigationTabUI(id, duration_options = c("1 HOUR", "24 HOUR"))
}
\arguments{
\item{id}{string ID the connects timeseriesInvestigationTabUI() to timeseriesInvestigationTabServer()}

\item{duration_options}{vector The durations that should be avilable for the second
parameter. A subset of '1Hr', '24Hr' and '8Hr'. Default
is c('1Hr', '24Hr').}
}
\description{
Used alongside timeseriesInvestigationTabServer() to add a collection of
reactive elements to the app, including multiple parameter selections,
a timeseries plot, a scatter plot, and a pollution rose. Wrapper for standalone
functions scatterServer(), timeseriesServer() and pollutionRoseServer().
}
