% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesUI.R
\name{timeseriesUI}
\alias{timeseriesUI}
\title{Timeseries UI Function}
\usage{
timeseriesUI(id)
}
\arguments{
\item{id}{string ID label that links timeseriesUI() to timeseriesServer()}
}
\description{
Used alongside timeseriesServer() to add a timeseries of 1hr and optionally 24hr
data (wrapped in a fluidRow) to an app. Timeseries will include zoom
functionality, where a box can be dragged across the plot area and
double-clicked. To reset zoom, double-click the plot area again.
}
