% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_to_expression.R
\name{unit_to_expression}
\alias{unit_to_expression}
\title{Convert units to formatted expression}
\usage{
unit_to_expression(unit)
}
\arguments{
\item{unit}{character vector A string or vector of unit values to be
converted}
}
\value{
character vector Formatted expressions
}
\description{
Convert units to formatted expression
}
