% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSelectHeaderUI.R
\name{mapSelectHeaderUI}
\alias{mapSelectHeaderUI}
\title{Header table for selected map site}
\usage{
mapSelectHeaderUI(labelId)
}
\arguments{
\item{labelId}{string ID to match with a call to output$<labelId> in the
server function of the app.}
}
\value{
object Collection of shiny UI elements
}
\description{
Displays specified columns from data passed to sitemapServer function
for site clicked on map.
}
